+++ rxads/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.shaishavgandhi.sample" >
+    package="com.shaishavgandhi.rxads" >
 
     <uses-permission android:name="android.permission.INTERNET" />
 </manifest>
